+++ Octokit/Clients/IIssueTimelineClient.cs
@@ -20,7 +20,7 @@ namespace Octokit
         /// <param name="owner">The owner of the repository</param>
         /// <param name="repo">The name of the repository</param>
         /// <param name="issueNumber">The issue number</param>
-        Task<IReadOnlyList<TimelineEventInfo>> GetAllForIssue(string owner, string repo, int issueNumber);
+        Task<IReadOnlyList<TimelineEventInfo>> GetAllForIssue(string owner, string repo, long issueNumber);
 
         /// <summary>
         /// Gets all the various events that have occurred around an issue or pull request.
@@ -32,7 +32,7 @@ namespace Octokit
         /// <param name="repo">The name of the repository</param>
         /// <param name="issueNumber">The issue number</param>
         /// <param name="options">Options for changing the API repsonse</param>
-        Task<IReadOnlyList<TimelineEventInfo>> GetAllForIssue(string owner, string repo, int issueNumber, ApiOptions options);
+        Task<IReadOnlyList<TimelineEventInfo>> GetAllForIssue(string owner, string repo, long issueNumber, ApiOptions options);
 
         /// <summary>
         /// Gets all the various events that have occurred around an issue or pull request.
@@ -42,7 +42,7 @@ namespace Octokit
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="issueNumber">The issue number</param>
-        Task<IReadOnlyList<TimelineEventInfo>> GetAllForIssue(long repositoryId, int issueNumber);
+        Task<IReadOnlyList<TimelineEventInfo>> GetAllForIssue(long repositoryId, long issueNumber);
 
         /// <summary>
         /// Gets all the various events that have occurred around an issue or pull request.
@@ -53,6 +53,6 @@ namespace Octokit
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="issueNumber">The issue number</param>
         /// <param name="options">Options for changing the API response</param>
-        Task<IReadOnlyList<TimelineEventInfo>> GetAllForIssue(long repositoryId, int issueNumber, ApiOptions options);
+        Task<IReadOnlyList<TimelineEventInfo>> GetAllForIssue(long repositoryId, long issueNumber, ApiOptions options);
     }
 }
